requirements.txt
fastapi
python-multipart
jinja2
uvicorn
pyngrok
python-dotenv
openai
pydub